+++ package.json
@@ -5,6 +5,7 @@
   "description": "A React component library to power all Palmetto UI.",
   "private": false,
   "main": "dist/index.js",
+  "module": "dist/index.esm.js",
   "types": "dist/src/index.d.ts",
   "files": [
     "dist"
@@ -75,6 +76,11 @@
     "@babel/plugin-transform-runtime": "^7.10.4",
     "@babel/polyfill": "^7.10.4",
     "@babel/preset-react": "^7.10.4",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-commonjs": "^18.0.0",
+    "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^11.2.1",
+    "@rollup/plugin-typescript": "^8.2.1",
     "@semantic-release/commit-analyzer": "^8.0.1",
     "@storybook/addon-a11y": "^6.0.27",
     "@storybook/addon-actions": "^6.0.27",
@@ -128,6 +134,7 @@
     "mutationobserver-shim": "^0.3.7",
     "node-sass": "^4.14.1",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "postcss": "^8.2.10",
     "postcss-loader": "^3.0.0",
     "pre-commit": "^1.2.2",
     "prop-types": "^15.7.2",
@@ -138,6 +145,12 @@
     "react-router-dom": "^5.2.0",
     "react-select-event": "^4.1.4",
     "resolve-url-loader": "^3.1.2",
+    "rollup": "^2.45.2",
+    "rollup-plugin-cleaner": "^1.0.0",
+    "rollup-plugin-copy": "^3.4.0",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-postcss": "^4.0.0",
+    "rollup-plugin-typescript2": "^0.30.0",
     "sass": "^1.26.8",
     "sass-loader": "^8.0.2",
     "sass-resources-loader": "^2.0.3",
