+++ src/components/Formik/FormikToggle/FormikToggle.tsx
@@ -1,3 +1,6 @@
+/* eslint-disable */
+// @ts-nocheck
+
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Toggle } from '../../Toggle/Toggle';
